requirements.txt
environs==9.5.0
python-telegram-bot==13.13
google-cloud-dialogflow==2.14.1
vk-api==11.9.9
notifiers==1.3.3
